# docs/specs/api_contract.yaml
openapi: 3.1.0
info:
  title: Sistema de Gestión de Préstamos TI API
  version: 1.0.0
  description: API REST para la gestión de inventario y préstamos de equipos TI.
servers:
  - url: /
    description: Servidor principal
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RolUsuario:
      type: string
      enum:
        - colaborador
        - administrador_ti
    EstadoUsuario:
      type: string
      enum:
        - activo
        - inactivo
    EstadoEquipo:
      type: string
      enum:
        - disponible
        - prestado
        - en_mantencion
        - de_baja
...
    Usuario:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nombre_completo:
          type: string
        rol:
          $ref: '#/components/schemas/RolUsuario'
        estado:
          $ref: '#/components/schemas/EstadoUsuario'
        creado_en:
          type: string
          format: date-time
      required:
        - id
        - email
        - nombre_completo
        - rol
        - estado
        - creado_en
    Categoria:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        descripcion:
          type:
            - string
            - 'null'
      required:
        - id
        - nombre
    Equipo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codigo_inventario:
          type: string
        categoria_id:
          type: string
          format: uuid
        categoria:
          $ref: '#/components/schemas/Categoria'
        marca:
          type: string
        modelo:
          type: string
        numero_serie:
          type: string
        estado:
          $ref: '#/components/schemas/EstadoEquipo'
        creado_en:
          type: string
          format: date-time
      required:
        - id
        - codigo_inventario
        - categoria_id
        - marca
        - modelo
        - numero_serie
        - estado
        - creado_en
    CreateEquipoDto:
      type: object
      properties:
        codigo_inventario:
          type: string
        categoria_id:
          type: string
          format: uuid
        marca:
          type: string
        modelo:
          type: string
        numero_serie:
          type: string
        estado:
          $ref: '#/components/schemas/EstadoEquipo'
      required:
        - codigo_inventario
        - categoria_id
        - marca
        - modelo
        - numero_serie
...
    SsoLoginDto:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    AuthResponseDto:
      type: object
      properties:
        access_token:
          type: string
        usuario:
          $ref: '#/components/schemas/Usuario'
      required:
        - access_token
        - usuario
    MetricasResumen:
      type: object
      properties:
        total_equipos:
          type: integer
        equipos_disponibles:
          type: integer
        equipos_prestados:
          type: integer
        equipos_en_mantencion:
          type: integer
        prestamos_activos:
          type: integer
        prestamos_vencidos:
          type: integer
        solicitudes_pendientes:
          type: integer
      required:
        - total_equipos
        - equipos_disponibles
        - equipos_prestados
        - equipos_en_mantencion
        - prestamos_activos
        - prestamos_vencidos
        - solicitudes_pendientes
...
paths:
  /api/v1/auth/sso-login:
    post:
      summary: Autenticación mediante SSO con Microsoft Entra ID
      operationId: ssoLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsoLoginDto'
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Token no válido o no autorizado
  /api/v1/auth/me:
    get:
      summary: Obtener perfil del usuario autenticado
      operationId: getAuthMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Datos del usuario autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '401':
          description: No autorizado
  /api/v1/categorias:
    get:
      summary: Listar todas las categorías de equipos
      operationId: getCategorias
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de categorías
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '401':
          description: No autorizado
  /api/v1/equipos:
    get:
      summary: Listar inventario de equipos con filtros
      operationId: getEquipos
      security:
        - bearerAuth: []
      parameters:
        - name: categoria_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: estado
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EstadoEquipo'
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista de equipos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipo'
        '401':
          description: No autorizado
    post:
      summary: Registrar nuevo equipo en inventario
      operationId: createEquipo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEquipoDto'
      responses:
        '201':
          description: Equipo creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipo'
        '400':
          description: Datos de entrada inválidos
        '401':
          description: No autorizado
        '403':
          description: Permisos insuficientes (requiere administrador_ti)
  /api/v1/equipos/{id}:
    get:
      summary: Obtener detalle de un equipo específico
      operationId: getEquipoById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalle del equipo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipo'
        '401':
          description: No autorizado
        '404':
          description: Equipo no encontrado
...
  /api/v1/metricas/resumen:
    get:
      summary: Obtener resumen numérico de métricas operacionales
      operationId: getMetricasResumen
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resumen de indicadores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricasResumen'
        '401':
          description: No autorizado
        '403':
          description: Permisos insuficientes (requiere administrador_ti)
...